Shift arenas 2M into center
This is to provide more space for the sound / lighting desk to be in the
top right of the plan, and space for lighting rig material / blah. This
also gives more space for people to be in the staging area.

This also moves the router to the inside of the crowd control barriers
-- this was suggested by Rob, to avoid the routing equipment being caked
in a layer of competitors.

Change-Id: I0e45357b348877bfd17f763c4f33b25e4117b3fb