Place plasma screens
These have gone in the gap between the pits and arenas because there'll
be a reasonable amount of space for people to stand around and look at
them, and it's somewhere that people can casually drift through.

Seeing how we mostly have 5 port switches and everywhere is saturated,
I've run both of the plasma screen network runs from the head pit
switch. It's likely to be better shielded from competitors, so will be
better than hanging the screens off a pit-router that competitors can
mess with. Plus, the head switch is already going to have to be a
biggie.

Change-Id: I7c38ae1200fc8d535e3900d1367ad2d5651e6f2b