Explicitly mark out switches in pits
Don't merge them with APs any more. Note that because there are 4 people
on each pit cluster and only 5 ports on switches, we need to run more
cable to hook up APs. To avoid gratuitous amounts of cable, put another
switch on two desks and run cable between it and the APs

Change-Id: Id847ee8744bf70ee32ff7fff9b84d7feea5f3ee3