Add power points and "arena" layer
List the number of power plugs we need at a particular location in the
arenas. Hopefully should provide a reasonable amount of information to
plan where stuff is allocated.

Also move the large blocks of arena colour to their own layer, so that
we can turn them off at will.

Change-Id: Icb8ba23a92fa8b5875deefb1f719c6ff74d3660e