Arrange this year's teams
The previous content of teams.txt was a copy of the teams from
SR2015, albeit reversed to have those on the upper floor first.
Since we've actually got quite a different list of teams competing
this year, import the proper (current) list of teams.

The new content maintains the same principle that teams which
were on the ground floor last year should be upstairs this year.

Change-Id: I0d147cdd5adae8e0cd8ec428698d8b2e844a2d08