Move and rename the displays for 2016
We've now got quite a few which aren't directly arena displays, so
clarify that in the legends.

Add the display on the top floor which was present last year but
not marked. Remove the second display from the ground floor which
was never deployed (and won't be). Add the two new screens in the
staging area (for use by the shepherds).

Change-Id: Ic5c6cf4421f11db39dc09b69e412e6817c771ad4